% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spectral.R
\name{apply_to_collection}
\alias{apply_to_collection}
\title{Apply function to collection of spectra}
\usage{
apply_to_collection(spectra_collection, fun, ...)
}
\arguments{
\item{spectra_collection}{A ruffs_spectra_collection object}

\item{fun}{Function to apply to each spectrum}

\item{...}{Additional arguments passed to fun}
}
\value{
A ruffs_spectra_collection with results
}
\description{
Apply function to collection of spectra
}
\examples{
\dontrun{
interpolated <- apply_to_collection(specs, interpolation, 
                                     target_wavelengths = seq(350, 2400, 1))
}
}
