% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-jump.R
\name{.get_segment_id}
\alias{.get_segment_id}
\title{Assign wavelengths to segments based on splice points}
\usage{
.get_segment_id(wavelength, splices)
}
\arguments{
\item{wavelength}{Numeric vector of wavelengths}

\item{splices}{Numeric vector of splice points}
}
\value{
Integer vector of segment IDs (1-based)
}
\description{
Assign wavelengths to segments based on splice points
}
\keyword{internal}
