% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-jump.R
\name{.jump_correct_additive}
\alias{.jump_correct_additive}
\title{Internal function for additive jump correction}
\usage{
.jump_correct_additive(wavelength, measurement, splices, reference)
}
\arguments{
\item{wavelength}{Numeric vector of wavelengths}

\item{measurement}{Numeric vector of measurements}

\item{splices}{Numeric vector of splice points}

\item{reference}{Integer indicating reference segment (1-based)}
}
\value{
Numeric vector of corrected measurements
}
\description{
Internal function for additive jump correction
}
\keyword{internal}
