% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-group.R
\name{get_group_names}
\alias{get_group_names}
\title{Get Group Names from Grouped Spectra}
\usage{
get_group_names(spectra)
}
\arguments{
\item{spectra}{A \code{ruffs_spectra_collection_grouped} object.}
}
\value{
Character vector of group names.
}
\description{
Retrieves the unique group names from a grouped spectral collection.
}
