% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spectral.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract metadata from RUFFS spectra}
\usage{
get_metadata(spectra, field = NULL)
}
\arguments{
\item{spectra}{A ruffs_spectra object}

\item{field}{Optional specific metadata field to extract}
}
\value{
List of all metadata, or specific field value
}
\description{
Extract metadata from RUFFS spectra
}
\examples{
\dontrun{
spec <- asdreader("test.asd")
get_metadata(spec)
get_metadata(spec, "instrument_type")
}
}
