% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-group.R
\name{get_spectra_by_group}
\alias{get_spectra_by_group}
\title{Get Spectra by Group}
\usage{
get_spectra_by_group(spectra, group_name)
}
\arguments{
\item{spectra}{A \code{ruffs_spectra_collection_grouped} object.}

\item{group_name}{Character string specifying which group to extract.}
}
\value{
A \code{ruffs_spectra_collection} object containing only spectra from the
specified group.
}
\description{
Extracts all spectra belonging to a specific group.
}
