% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-group.R
\name{list_groups}
\alias{list_groups}
\title{List Spectra in Each Group}
\usage{
list_groups(spectra)
}
\arguments{
\item{spectra}{A \code{ruffs_spectra_collection_grouped} object.}
}
\value{
A named list where names are group names and values are
character vectors of spectrum filenames.
}
\description{
Returns a list showing which spectra belong to each group.
}
