% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spectral.R
\name{new_ruffs_spectra_collection}
\alias{new_ruffs_spectra_collection}
\title{Create a RUFFS spectra collection}
\usage{
new_ruffs_spectra_collection(spectra_list)
}
\arguments{
\item{spectra_list}{List of ruffs_spectra objects}
}
\value{
A ruffs_spectra_collection object
}
\description{
Constructor for ruffs_spectra_collection class - a list of ruffs_spectra objects
}
\keyword{internal}
