% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spectral.R
\name{print.ruffs_spectra_collection}
\alias{print.ruffs_spectra_collection}
\title{Print method for ruffs_spectra_collection}
\usage{
\method{print}{ruffs_spectra_collection}(x, ...)
}
\description{
Print method for ruffs_spectra_collection
}
