% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-group.R
\name{print.ruffs_spectra_collection_grouped}
\alias{print.ruffs_spectra_collection_grouped}
\title{Print Method for Grouped Spectra Collection}
\usage{
\method{print}{ruffs_spectra_collection_grouped}(x, ...)
}
\arguments{
\item{x}{A \code{ruffs_spectra_collection_grouped} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Grouped Spectra Collection
}
