% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-summarize.R
\name{print.ruffs_spectra_summary}
\alias{print.ruffs_spectra_summary}
\title{Print Method for Spectra Summary}
\usage{
\method{print}{ruffs_spectra_summary}(x, ...)
}
\arguments{
\item{x}{A \code{ruffs_spectra_summary} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Spectra Summary
}
