% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spectral.R
\name{[.ruffs_spectra}
\alias{[.ruffs_spectra}
\title{Subset method for ruffs_spectra that preserves attributes}
\usage{
\method{[}{ruffs_spectra}(x, i, j, drop = FALSE)
}
\description{
Subset method for ruffs_spectra that preserves attributes
}
